+++ packages/core/utils/common.ts
@@ -1,6 +1,5 @@
 import * as types from './types';
 import { dispatchToMainThread, dispatchToUIThread, isMainThread } from './mainthread-helper';
-import { sanitizeModuleName } from '../ui/builder/module-name-sanitizer';
 import emojiRegex from 'emoji-regex';
 
 import { GC } from './index';
@@ -41,6 +40,31 @@ export function getModuleName(path: string): string {
 	const moduleName = path.replace('./', '');
 
 	return sanitizeModuleName(moduleName);
+}
+
+/**
+ * Helps sanitize a module name if it is prefixed with '~/', '~' or '/'
+ * @param moduleName the name
+ * @param removeExtension whether to remove extension
+ */
+export function sanitizeModuleName(moduleName: string, removeExtension = true): string {
+	moduleName = moduleName.trim();
+
+	if (moduleName.startsWith('~/')) {
+		moduleName = moduleName.substring(2);
+	} else if (moduleName.startsWith('~')) {
+		moduleName = moduleName.substring(1);
+	} else if (moduleName.startsWith('/')) {
+		moduleName = moduleName.substring(1);
+	}
+
+	if (removeExtension) {
+		const extToRemove = ['js', 'ts', 'xml', 'html', 'css', 'scss'];
+		const extensionRegEx = new RegExp(`(.*)\\.(?:${extToRemove.join('|')})`, 'i');
+		moduleName = moduleName.replace(extensionRegEx, '$1');
+	}
+
+	return moduleName;
 }
 
 export function isFileOrResourcePath(path: string): boolean {
@@ -209,4 +233,56 @@ export function isEmoji(value: string): boolean {
 	// TODO: In a future runtime update, we can switch to using Unicode Property Escapes:
 	// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions/Unicode_Property_Escapes
 	return emojiRegex().test(value);
+}
+
+/**
+ * Default animation values used throughout core
+ */
+export const CORE_ANIMATION_DEFAULTS = {
+	duration: 0.35,
+	spring: {
+		tension: 140,
+		friction: 10,
+		mass: 1,
+		velocity: 0,
+	},
+};
+
+/**
+ * Get a duration with damping value from various spring related settings.
+ * Helpful when needing to convert spring settings to isolated duration value.
+ * @param springSettings various spring settings
+ * @returns calculated duration with damping from spring settings
+ */
+export function getDurationWithDampingFromSpring(springSettings?: { tension?: number; friction?: number; mass?: number; velocity?: number }) {
+	// for convenience, default spring settings are provided
+	const opt = {
+		...CORE_ANIMATION_DEFAULTS.spring,
+		...(springSettings || {}),
+	};
+	const damping = opt.friction / Math.sqrt(2 * opt.tension);
+	const undampedFrequency = Math.sqrt(opt.tension / opt.mass);
+
+	// console.log({
+	// 	damping,
+	// 	undampedFrequency
+	// })
+
+	const epsilon = 0.001;
+	let duration = 0;
+
+	if (damping < 1) {
+		// console.log('damping < 1');
+		const a = Math.sqrt(1 - Math.pow(damping, 2));
+		const b = opt.velocity / (a * undampedFrequency);
+		const c = damping / a;
+		const d = -((b - c) / epsilon);
+		if (d > 0) {
+			duration = Math.log(d) / (damping * undampedFrequency);
+		}
+	}
+	return {
+		duration,
+		damping,
+	};
 }
